]仕様追加]パルス幅取得処理の完成版をコミット

+++ PIC18F25Q10_MRC_program.X/source/userdefine.h
@@ -1,6 +1,6 @@
 /* 
- * File:   adc.h
- * Author: shunt
+ * File:   userdefine.h
+ * Author: ICE_MEGANE
  *
  * Created on 2026/03/04, 22:26
  */
@@ -23,21 +23,20 @@ extern "C" {
 #define     ON      1U
 #define     HI      1U
 
-
 /* 共通型名称 */
 typedef  unsigned char   u8;
 typedef  unsigned short  u16;
 typedef  unsigned long   u32;
 
-
 #define     U8_MAX     ((u8)0xFF)
-#define     U16_MAX     ((u8)0xFFFF)
-#define     U32_MAX     ((u8)0xFFFFFFFF)
+#define     U16_MAX    ((u16)0xFFFF)
+#define     U32_MAX    ((u32)0xFFFFFFFF)
+
 
 
 /* 全ファイル共通関数 */
 extern u32 func_ud_g_calcmul_2x2_byte( u16 u16_arg1, u16 u16_arg2 );
-extern u32 func_ud_g_calcdiv_4x4_byte( u32 u32_arg1, u16 u32_arg2 );
+extern u32 func_ud_g_calcdiv_4x4_byte( u32 u32_arg1, u32 u32_arg2 );
 
 #ifdef	__cplusplus
 }
